更新NGUI版本至3.7.3

+++ ClientFramework/Assets/NGUI/Scripts/Internal/Localization.cs
@@ -29,6 +29,14 @@ using System.Collections.Generic;
 
 public static class Localization
 {
+	public delegate byte[] LoadFunction (string path);
+
+	/// <summary>
+	/// Want to have Localization loading be custom instead of just Resources.Load? Set this function.
+	/// </summary>
+
+	static public LoadFunction loadFunction;
+
 	/// <summary>
 	/// Whether the localization dictionary has been loaded.
 	/// </summary>
@@ -115,21 +123,36 @@ public static class Localization
 	static bool LoadDictionary (string value)
 	{
 		// Try to load the Localization CSV
-		TextAsset txt = localizationHasBeenSet ? null : Resources.Load("Localization", typeof(TextAsset)) as TextAsset;
-		localizationHasBeenSet = true;
+		byte[] bytes = null;
+
+		if (!localizationHasBeenSet)
+		{
+			if (loadFunction == null)
+			{
+				TextAsset asset = Resources.Load<TextAsset>("Localization");
+				if (asset != null) bytes = asset.bytes;
+			}
+			else bytes = loadFunction("Localization");
+			localizationHasBeenSet = true;
+		}
 
 		// Try to load the localization file
-		if (txt != null && LoadCSV(txt)) return true;
+		if (LoadCSV(bytes)) return true;
 
 		// If this point was reached, the localization file was not present
 		if (string.IsNullOrEmpty(value)) return false;
 
 		// Not a referenced asset -- try to load it dynamically
-		txt = Resources.Load(value, typeof(TextAsset)) as TextAsset;
-
-		if (txt != null)
+		if (loadFunction == null)
 		{
-			Load(txt);
+			TextAsset asset = Resources.Load<TextAsset>(value);
+			if (asset != null) bytes = asset.bytes;
+		}
+		else bytes = loadFunction(value);
+
+		if (bytes != null)
+		{
+			Set(value, bytes);
 			return true;
 		}
 		return false;
@@ -147,8 +170,12 @@ public static class Localization
 			if (SelectLanguage(value)) return true;
 		}
 
+		// Old style dictionary
+		if (mOldDictionary.Count > 0) return true;
+
 		// Either the language is null, or it wasn't found
 		mOldDictionary.Clear();
+		mDictionary.Clear();
 		if (string.IsNullOrEmpty(value)) PlayerPrefs.DeleteKey("Language");
 		return false;
 	}
@@ -164,12 +191,35 @@ public static class Localization
 	}
 
 	/// <summary>
+	/// Set the localization data directly.
+	/// </summary>
+
+	static public void Set (string languageName, byte[] bytes)
+	{
+		ByteReader reader = new ByteReader(bytes);
+		Set(languageName, reader.ReadDictionary());
+	}
+
+	/// <summary>
 	/// Load the specified CSV file.
 	/// </summary>
 
-	static public bool LoadCSV (TextAsset asset)
+	static public bool LoadCSV (TextAsset asset) { return LoadCSV(asset.bytes, asset); }
+
+	/// <summary>
+	/// Load the specified CSV file.
+	/// </summary>
+
+	static public bool LoadCSV (byte[] bytes) { return LoadCSV(bytes, null); }
+
+	/// <summary>
+	/// Load the specified CSV file.
+	/// </summary>
+
+	static bool LoadCSV (byte[] bytes, TextAsset asset)
 	{
-		ByteReader reader = new ByteReader(asset);
+		if (bytes == null) return false;
+		ByteReader reader = new ByteReader(bytes);
 
 		// The first line should contain "KEY", followed by languages.
 		BetterList<string> temp = reader.ReadCSV();
@@ -246,7 +296,15 @@ public static class Localization
 		if (values.size < 2) return;
 		string[] temp = new string[values.size - 1];
 		for (int i = 1; i < values.size; ++i) temp[i - 1] = values[i];
-		mDictionary.Add(values[0], temp);
+		
+		try
+		{
+			mDictionary.Add(values[0], temp);
+		}
+		catch (System.Exception ex)
+		{
+			Debug.LogError("Unable to add '" + values[0] + "' to the Localization dictionary.\n" + ex.Message);
+		}
 	}
 
 	/// <summary>
@@ -299,8 +357,13 @@ public static class Localization
 	}
 
 	/// <summary>
-	/// Localize the specified value.
+	/// Localize the specified value and format it.
 	/// </summary>
+
+	static public string Format (string key, params object[] parameters) { return string.Format(Get(key), parameters); }
+
+	[System.Obsolete("Localization is now always active. You no longer need to check this property.")]
+	static public bool isActive { get { return true; } }
 
 	[System.Obsolete("Use Localization.Get instead")]
 	static public string Localize (string key) { return Get(key); }
@@ -311,7 +374,14 @@ public static class Localization
 
 	static public bool Exists (string key)
 	{
-		if (mLanguageIndex != -1) return mDictionary.ContainsKey(key);
-		return mOldDictionary.ContainsKey(key);
+		// Ensure we have a language to work with
+		if (!localizationHasBeenSet) language = PlayerPrefs.GetString("Language", "English");
+
+#if UNITY_IPHONE || UNITY_ANDROID
+		string mobKey = key + " Mobile";
+		if (mDictionary.ContainsKey(mobKey)) return true;
+		else if (mOldDictionary.ContainsKey(mobKey)) return true;
+#endif
+		return mDictionary.ContainsKey(key) || mOldDictionary.ContainsKey(key);
 	}
 }
